Cloud Computing/Anabul Breed Detection Backend/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
requests